角色增加layui适配 以及人员角色更新时的bug（第一次设置人员没有旧角色所报的错）

+++ workflow/src/main/java/com/xonro/vflow/workflow/service/impl/UserServiceImpl.java
@@ -274,7 +274,10 @@ public class UserServiceImpl implements UserService {
             }
             //删除旧角色
             Group group = identityService.createGroupQuery().groupType(OrgEnum.GROUP_TYPE_ROLE.getValue()).groupMember(userId).singleResult();
-            identityService.deleteMembership(userId,group.getId());
+            if (group!=null){
+                identityService.deleteMembership(userId,group.getId());
+            }
+
 
             //创建新角色
             identityService.createMembership(userId,role.getGroupId());
